+++ app/src/main/java/com/dsvag/androidacademyproject/ui/person/PersonViewModel.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.dsvag.androidacademyproject.data.repositories.PersonRepository
-import com.dsvag.androidacademyproject.models.movies.Movie
+import com.dsvag.androidacademyproject.models.movie.Movie
 import com.dsvag.androidacademyproject.models.person.Person
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
@@ -19,7 +19,7 @@ class PersonViewModel @ViewModelInject constructor(
     private val _mutablePersonMovieData: MutableLiveData<List<Movie>> = MutableLiveData()
     val personMovieData get() = _mutablePersonMovieData
 
-    fun fetchPerson(personId: Int) {
+    fun fetchPerson(personId: Long) {
         viewModelScope.launch(Dispatchers.IO) {
             val person = runCatching { personRepository.getPerson(personId) }.getOrNull()
 
@@ -29,7 +29,7 @@ class PersonViewModel @ViewModelInject constructor(
         }
     }
 
-    fun fetchPersonMovie(personId: Int) {
+    fun fetchPersonMovie(personId: Long) {
         viewModelScope.launch(Dispatchers.IO) {
             val personMovies =
                 runCatching { personRepository.getPersonMovies(personId) }.getOrNull()
